[파일 수정] 충돌 수정, laundry 수정중

+++ src/main/java/com/laundry_m/mvc/dao/LaundryDaoImpl.java
@@ -221,7 +221,7 @@ public class LaundryDaoImpl implements LaundryDao {
 	}
 
 	@Override
-	public Laundry selectByLowestByLaundry(Long clothesId, Long FabricId) throws SQLException {
+	public Laundry selectByLowestByLaundry(int clothesId, int FabricId) throws SQLException {
 		SqlSession session = null;
 		Laundry laundry = null;
